Failed attempts
Well, everything looks like it's working! It's not actually working, but sure doesn't look like there are any problems. Hate my life...

So, when a tab is "ready" it attaches a content script file which searches for scripts inside that tab, and disables them. Each of those scripts is emitted to the main worker, which hashes the source, and conditionally re-enables the script. The third worker for the panel fills the table rows on that panel, and when a check box is toggled it emits to the main worker to change the allowed setting for a certain script.

When the widget is clicked in the main worker, it emits to the panel worker to fill itself up with table rows, getting those rows from what the tab ready worker has already reported to main, according to whatever tab is the active one. That way the panel will only show info about the scripts for that page. That means main has to listen to the tab ready worker and keep track of all scripts on all tabs, to produce on demand whenever the widget is clicked.

I don't know what's supposed to happen if you click the widget before the active tab has loaded. It's probably another unavoidable race condition. I also don't know whether the tab ready worker even manages to get rid of the page's scripts before they are evaluated. It probably only gets to remove script elements after they've all been run, thus defeating the purpose. I don't know why when the tab becomes ready, the tab ready worker does nothing where it should be listing and extracting scripts from whatever page is loading in that tab. I am basically the worst person possible to make an add-on like this. I have no idea what I'm doing, I'm horribly insecure, and not very intelligent. But since everyone else in the world seems to think that trusting javascript according to whatever host the DNS name resolves to, and whatever happens to be on the host at that URL today, is not a hideous security risk, I'm the only one who even tried. For shame!